requirements.txt
pyyaml==6.0.1
fastapi==0.111.0
langchain==0.2.6
langchain_openai==0.1.13
langgraph==0.1.5
cryptography==43.0.0
haystack_ai==2.3.1
pypdf==4.3.1
python-docx==1.1.2
pymysql==1.1.1
pymysql-pool==0.4.6
pydantic==2.8.2
#pygraphviz==1.13 # Only for analysis
elasticsearch-haystack==0.7.1
jieba==0.42.1
sqlalchemy==2.0.34
datasets==3.0.0
nest_asyncio==1.6.0
ragas==0.1.18
xinference_client==0.15.0